+++ XGrid/Window controllers/GridLines window controller/GridLinesPreferences.h
@@ -5,7 +5,7 @@
 
 @property (nonatomic) LinesSettings *linesSettings;
 
-@property (nonatomic, copy) void (^onSliderChange)(uint8);
+@property (nonatomic, copy) void (^onSliderChange)(CGFloat);
 @property (nonatomic, copy) void (^onColorSelected)(NSColor*);
 @property (nonatomic, copy) void (^willClose)(LinesSettings*);
