+++ XGrid/Window controllers/GridLines window controller/View controller/Row views/ColorRowView/ColorRowView.m
@@ -12,7 +12,7 @@
 #import "ColorCircleItem.h"
 
 @interface ColorRowView () <NSCollectionViewDelegate, NSCollectionViewDataSource, NSCollectionViewDelegateFlowLayout> {
-    uint8 size;
+    CGFloat size;
     NSScrollView *scrollView;
     NSCollectionView *collectionView;
     NSString *reuseColorItem;
